+++ src/components/Cards/Cards.module.css
@@ -111,8 +111,6 @@
    font-weight: 400;
 }
 .chalengeName{
-  
-  
    font-size: 16px;
    font-weight: 700;
    display: flex;
@@ -122,7 +120,6 @@
    margin: 0;
 }
 .taskName{
-   width: 90%;
    width: 90%;
    display: flex;
    justify-content: center;
