+++ src/components/CardsList/CardsList.jsx
@@ -1,11 +1,11 @@
-import React, { useState } from 'react';
-import Card from "../Cards/Cards"
+import React from 'react';
 import style from "./CardsList.module.css"
 import Clock from "./Clock"
+import TodayCards from './TodayCards';
+import TomorrowCards from './TomorrowCards';
 
 function CardsList({cards}) {
 
- 
    const addCard = () => {
      const newCard = {
        isChallenge: false,
@@ -15,94 +15,13 @@ function CardsList({cards}) {
      };
      setCards([...cards, newCard]);
    };
-//
-  const currentDate = new Date();
-  const tomorrowDate = new Date(currentDate.getTime() + 24 * 60 * 60 * 1000);
-  
-  const filteredCards = cards.filter((card) => {
-    const cardDate = new Date(card.date);
-    return (
-      cardDate.getFullYear() === currentDate.getFullYear() &&
-      cardDate.getMonth() === currentDate.getMonth() &&
-      cardDate.getDate() === currentDate.getDate()
-    );
-  });
-  const todayCards = filteredCards.sort((a, b) => {
-   const timeA = timeStringToMs(a.time);
-		const timeB = timeStringToMs(b.time);
-      console.log(timeA)
-      console.log(timeB)
-		if (timeA < timeB) {
-			return -1;
-		} else if (timeA > timeB) {
-			return 1;
-		} else {
-			return 0;
-		}
-  });
-  const filteredTomorrowCards = cards.filter((card) => {
-    const cardDate = new Date(card.date);
-    return (
-      cardDate.getFullYear() === tomorrowDate.getFullYear() &&
-      cardDate.getMonth() === tomorrowDate.getMonth() &&
-      cardDate.getDate() === tomorrowDate.getDate()
-    );
-  });
- 
-const tomorrowCards = filteredTomorrowCards.sort((a, b) => {
-		const timeA = timeStringToMs(a.time);
-		const timeB = timeStringToMs(b.time);
-      console.log(timeA)
-      console.log(timeB)
-		if (timeA < timeB) {
-			return -1;
-		} else if (timeA > timeB) {
-			return 1;
-		} else {
-			return 0;
-		}
-	});
-function timeStringToMs(cards) {
-  const [hours, minutes] = cards.split("-");
-  const date = new Date();
-  date.setHours(hours);
-  date.setMinutes(minutes);
-  return date.getTime();
-}
 
    return(
       <div className={style.sectionCards}>
          <button onClick={addCard}>Add Card</button>
          <Clock/>
-         <div className={style.todaylist}>
-            <h4>Today</h4>
-            <ul className={style.cardList}>
-            {todayCards.map((card) => (
-               <Card key={card.id} 
-               title={card.title} 
-               dueDate={card.date} 
-               dueTime={card.time}
-               isChallenge={card.isChallenge}
-               difficulty={card.difficulty}
-               category={card.category}
-               /> ))}
-         </ul>
-         </div>
-         <div className={style.tomorrowList}>
-            <h4>Tomorrow </h4>
-               <ul className={style.cardList}>
-               {tomorrowCards.map((card) => (
-                  <Card key={card.id} 
-                  title={card.title} 
-                  dueDate={card.date} 
-                  dueTime={card.time}
-                  isChallenge={card.isChallenge}
-                  difficulty={card.difficulty}
-                  category={card.category}
-                  />
-                  ))}
-               </ul>
-         </div>
+         <TodayCards cards={cards}/>
+         <TomorrowCards cards={cards}/>
          
       </div>
    )
